refactor: organize header files
