+++ src/ui/header/stats.ui.ts
@@ -1,5 +1,5 @@
-import { UI_POSITIONS, INFO_MSGS } from '../constants/index.js';
-import { Ui } from './ui.js';
+import { UI_POSITIONS, INFO_MSGS } from '../../constants/index.js';
+import { Ui } from '../ui.js';
 import { ResultsService } from 'src/services/results.service.js';
 
 export class StatsUi extends Ui {
